+++ mobile/library/objective-c/EnvoyEngine.h
@@ -385,10 +385,13 @@ extern const int kEnvoyFailure;
  running.
  @param logger Logging interface.
  @param eventTracker Event tracking interface.
+ @param enableNetworkPathMonitor Configure the engine to use `NWPathMonitor` to observe network
+ reachability.
  */
 - (instancetype)initWithRunningCallback:(nullable void (^)())onEngineRunning
                                  logger:(nullable void (^)(NSString *))logger
-                           eventTracker:(nullable void (^)(EnvoyEvent *))eventTracker;
+                           eventTracker:(nullable void (^)(EnvoyEvent *))eventTracker
+               enableNetworkPathMonitor:(BOOL)enableNetworkPathMonitor;
 /**
  Run the Envoy engine with the provided configuration and log level.
 
@@ -518,9 +521,15 @@ extern const int kEnvoyFailure;
 // Monitors network changes in order to update Envoy network cluster preferences.
 @interface EnvoyNetworkMonitor : NSObject
 
-// Start monitoring reachability, updating the preferred Envoy network cluster on changes.
+// Start monitoring reachability using `SCNetworkReachability`, updating the
+// preferred Envoy network cluster on changes.
 // This is typically called by `EnvoyEngine` automatically on startup.
 + (void)startReachabilityIfNeeded;
+
+// Start monitoring reachability using `NWPathMonitor`, updating the
+// preferred Envoy network cluster on changes.
+// This is typically called by `EnvoyEngine` automatically on startup.
++ (void)startPathMonitorIfNeeded API_AVAILABLE(ios(12));
 
 @end